+++ database/migrations/2024_12_03_232652_create_customer_payments_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('customer_payments', function (Blueprint $table) {
             $table->id();
-            $table->float('value');
+            $table->integer('value');
             $table->foreignId(Sell::class);
             $table->timestamps();
         });
